+++ src/main/java/com/flightstats/hub/cluster/HubClusterRegister.java
@@ -1,16 +1,17 @@
-package com.flightstats.hub.spoke;
+package com.flightstats.hub.cluster;
 
 import com.flightstats.hub.app.HubServices;
-import com.flightstats.hub.cluster.CuratorCluster;
 import com.google.common.util.concurrent.AbstractIdleService;
 import com.google.inject.Inject;
+import com.google.inject.Singleton;
 import com.google.inject.name.Named;
 
+@Singleton
 public class HubClusterRegister {
 
     @Inject
     @Named("HubCuratorCluster")
-    private CuratorCluster spokeCuratorCluster;
+    private CuratorCluster hubCuratorCluster;
 
     public HubClusterRegister() {
         HubServices.register(new CuratorClusterHook(), HubServices.TYPE.FINAL_POST_START, HubServices.TYPE.PRE_STOP);
@@ -19,13 +20,13 @@ public class HubClusterRegister {
     private class CuratorClusterHook extends AbstractIdleService {
         @Override
         protected void startUp() throws Exception {
-            spokeCuratorCluster.addCacheListener();
-            spokeCuratorCluster.register();
+            hubCuratorCluster.addCacheListener();
+            hubCuratorCluster.register();
         }
 
         @Override
         protected void shutDown() throws Exception {
-            spokeCuratorCluster.delete();
+            hubCuratorCluster.delete();
         }
     }
 }
